warehouse/backends/requirements.txt
django
mysqlclient
djangorestframework
markdown
django-filter
